fhir resource and process sate management
Process state are tracked by the BPE and stored in the DB. FHIR
resources are created/updated or deleted accordingly.

Process state can be:
NEW -> Found in jars, not deployed yet
ACTIVE -> Found in jars, not excluded by property, not retired by
property, not draft by -SNAPSHOT.jar as end of jar name or -SNAPSHOT as
end of folder name
RETIRED -> Found in jars, retired by property
DRAFT -> Found in jars, jar name ends with -SNAPSHOT.jar or folder name
ends with -SNAPSHOT
EXCLUDED -> Not found in jars (aka jar deleted or jar renamed to e.g.
...jar.off), or excluded by property

FHIR resources for processes with status DRAFT will be updated in FHIR
server on every restart. FHIR resources for processes with status ACTIVE
will only be updated if the status changes to ACTIVE but not on restart
without status change. A status change to EXCLUDED will result in
deletions of resources that are not needed by other processes.