reworked autostart process with unique instances based on BPMN signals
The new ping autostart process implementation uses an event based
gateway to either wait for a timer to fire after 24 hours or a global
signal to abort.
The abort signal is send as the first step of the autostart process to
stop existing instances and based on a stopPingAutostart Task resource.
The new implementation does not require (nor allow) to specify a
business-key via Task resource.
If a stopPingAutostart Task resource is send while the autostart process
is active and not waiting, nothing happens. Stopping the autostart
process is only possible while the process is sleeping but then is
stopped immediately.